Reassemble frames split across multiple responses
We previously ignored the flags and offset fields in responses, and
treated each response as one frame. Now, we look at the size field,
allocate a buffer of that size and wait until it is filled.

We also now use the offset (previously we assumed it was always zero).

A side-effect of these changes is that listen now starts and returns the
polling thread. Before, it returned a dummy thread that never resolved.
This means that errors in the listen thread should now be reported.

Errors in the user function are now handled by Lwt.async (however that
is configured). Before, we always printed the exception and then
continued).

Another change is that we previously started listening immediately but
discarded all packets until listen was called. It's not clear why we
wanted to do that (though a comment indicated it was the intended
behaviour). Not dropping packets seems like a more useful behaviour and
simplifies the code. The user can always supply a function that drops
packets if they want that.